+++ src/Controller/Admin/BannersController.php
@@ -23,6 +23,7 @@
 namespace MeCms\Controller\Admin;
 
 use MeCms\Controller\AppController;
+use MeCms\Controller\Traits\DownloadTrait;
 
 /**
  * Banners controller
@@ -30,6 +31,8 @@ use MeCms\Controller\AppController;
  */
 class BannersController extends AppController
 {
+    use DownloadTrait;
+
     /**
      * Called before the controller action.
      * You can use this method to perform logic that needs to happen before
@@ -193,7 +196,7 @@ class BannersController extends AppController
      * Downloads banner
      * @param string $id Banner ID
      * @return \Cake\Network\Response
-     * @uses MeCms\Controller\AppController::_download()
+     * @uses MeCms\Controller\Traits\DownloadTrait::_download()
      */
     public function download($id = null)
     {